pyproject.toml
[tool.poetry]
name = "whisper-journal"
version = "0.1.0"
description = "AI-powered audio journaling with Whisper"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12.0"
openai-whisper = "^20240930"
python-dotenv = "^1.0.0"
llvmlite = "^0.43.0"
numba = "^0.60.0"
rich = "^13.7.0"
typer = "^0.9.0"
whisper-live = "^0.5.1"
onnxruntime = "=1.16.0"
sounddevice = "=0.5.1"
soundfile = "=0.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["setuptools>=75.3.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
journal = "ai_journal.cli:app"